+++ changelog.md
@@ -12,6 +12,9 @@ New effects:
 - New SAO: Druid's Lunar and Solar Eclipse (Cataclysm)
 - New SAO: Druid's Shooting Stars (Cataclysm)
 - New SAO: Hunter's Lock and Load (Season of Discovery)
+- New SAO: Warrior's Bloodsurge (Cataclysm)
+- New SAO: Warrior's Sudden Death (Cataclysm)
+- New SAO: Warrior's Sword and Board (Cataclysm)
 - New SAO: Warrior's Sword and Board (Season of Discovery)
 - New GAB: Death Knight's Blood Boil, during Crimson Scourge (Cataclysm)
 - New GAB: Death Knight's Death Coil, during Sudden Doom (Cataclysm)
@@ -27,8 +30,11 @@ New effects:
 - New GAB: Paladin's Divine Light, during Infusion of Light (Cataclysm)
 - New GAB: Paladin's Holy Radiance, during Infusion of Light (Cataclysm)
 - New GAB: Paladin's Exorcism, during The Art of War (Cataclysm)
+- New GAB: Warrior's Slam, during Bloodsurge (Cataclysm)
+- New GAB: Warrior's Colossus Smash, during Sudden Death (Cataclysm)
+- New GAB: Warrior's Shield Slam, during Sword and Board (Cataclysm)
+- New GAB: Warrior's Shield Slam, during Sword and Board (Season of Discovery)
 - New GAB: Warrior's Overpower, during Taste for Blood* (Season of Discovery)
-- New GAB: Warrior's Slam, during Sword and Board (Season of Discovery)
 
 * As in Wrath, there is no dedicated Taste for Blood option
 Taste for Blood shares its option with Overpower
@@ -46,6 +52,9 @@ Bug fixes
 - Death Knight's Killing Machine (Cataclysm) no longer glows Howling Blast
 - Druid's Lunar or Solar Eclipse spell alert is displayed upon login if needed
 - Hungering Cold is now tagged as Frozen debuff for Mages
+- Warrior's Blood Surge (Season of Discovery) did not trigger correctly
+- Warrior's Bloodsurge (Wrath) did not preview from the options panel
+- Warrior's Sudden Death (Wrath) did not preview from the options panel
 
 #### v1.3.1 (2024-04-22)